pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "astropy-hdf5io"
version = "0.1.0"
description = "HDF5 serialization support for AstroPy objects using fsc.hdf5-io"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "BSD-3-Clause"}
authors = [
    {name = "astropy-hdf5io contributors", email = "your.email@example.com"}
]
maintainers = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = [
    "astropy",
    "hdf5",
    "serialization",
    "astronomy",
    "astrophysics",
    "io",
    "data"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Astronomy",
    "Topic :: Scientific/Engineering :: Physics",
    "Operating System :: OS Independent",
]
dependencies = [
    "astropy>=5.0",
    "fsc.hdf5-io>=1.0",
    "h5py>=3.0",
    "numpy>=1.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]
docs = [
    "sphinx>=5.0",
    "sphinx-rtd-theme>=1.0",
    "sphinx-autoapi>=2.0",
    "myst-parser>=1.0",
]

[project.urls]
Homepage = "https://github.com/liamh/astropy-hdf5io"
Documentation = "https://astropy-hdf5io.readthedocs.io/"
Repository = "https://github.com/liamh/astropy-hdf5io.git"
Issues = "https://github.com/liamh/astropy-hdf5io/issues"
Changelog = "https://github.com/liamh/astropy-hdf5io/blob/main/CHANGELOG.md"

[tool.setuptools]
packages = ["astropy_hdf5io"]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]  # line too long (handled by black)

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false